refactor: auth context 통합
- null check 추가
