+++ src/hooks/useAuth.ts
@@ -5,6 +5,10 @@ export function useAuth() {
   const auth = useContext(AuthContext);
   const dispatch = useContext(AuthDispatchContext);
 
+  if (!auth || !dispatch) {
+    throw Error('Auth Context is null');
+  }
+
   return {
     auth,
     authDispatch: dispatch,
