+++ src/utils/type.ts
@@ -14,14 +14,6 @@ export interface CreateChannelContent {
   description: string;
 }
 
-export interface Auth {
-  isSignin: boolean;
-  channelId: string | null;
-  channelName: string | null;
-}
-
-export interface AuthAction { type: 'signin' | 'signout' }
-
 export interface AuthSession {
   channelToken: string | null;
   channelId: string | null;
